MAINT: Fix test_masks
