+++ docker/files/nipype.cfg
@@ -1,0 +1,7 @@
+[execution]
+hash_method = content
+poll_sleep_duration = 0.01
+remove_unnecessary_outputs = true
+crashfile_format = txt
+profile_runtime = false
+use_relative_paths = false
